[CCAP-915] - removes ENABLE_PROVIDER_REGISTRATION_WITH_ITIN
