+++ sample.env
@@ -121,5 +121,4 @@ ENABLE_DTS_INTEGRATION=false
 ENABLE_RESOURCE_ORG_EMAIL=false
 ALLOW_PROVIDER_REGISTRATION_FLOW=true
 ENABLE_PROVIDER_MESSAGING=true
-ENABLE_PROVIDER_REGISTRATION_WITH_ITIN=true
 ENABLE_PROVIDER_RESPONSE_FEIN=true
