+++ src/main/java/org/ilgcc/app/submission/conditions/RegistrationForInHomeCare.java
@@ -14,9 +14,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ProviderITINRegistrationForInHomeCare implements Condition {
-    // registration-home-provider-ssn when itin flag is off
-    // registration-home-provider-tax-id when itin flag is on
+public class RegistrationForInHomeCare implements Condition {
 
     final static List providerTypesRequired = List.of(
             LICENSED_DAY_CARE_HOME.name(),
@@ -30,12 +28,9 @@ public class ProviderITINRegistrationForInHomeCare implements Condition {
     @Value("${il-gcc.allow-provider-registration-flow}")
     private boolean enableProviderRegistration;
 
-    @Value("${il-gcc.enable-provider-registration-with-itin}")
-    private boolean enableProviderRegistrationWithITIN;
-
     @Override
     public Boolean run(Submission submission) {
-        return enableProviderRegistrationWithITIN && enableProviderRegistration && displayScreen(submission);
+        return enableProviderRegistration && displayScreen(submission);
     }
 
     private Boolean displayScreen(Submission submission) {
