+++ src/main/resources/messages.properties
@@ -1515,7 +1515,6 @@ registration-basic-info-intro.2=CCAP payment details
 # registration-home-provider-ssn
 registration-home-provider-ssn.title=Social Security Number (SSN)
 registration-home-provider-ssn.header=What is your Social Security Number?
-registration-home-provider-ssn.subtext=<p>Since you are providing child care in your home, we need your Social Security Number (SSN) to check your current work situation.</p><p>This helps us make sure you aren't working during the hours you are scheduled for child care.</p>
 registration-home-provider-ssn.secure-footer=Your information is secure and will be handled in accordance with our <a target="_blank" data-mixpanel="open-privacy-policy" rel="noopener noreferrer" href="/privacy">privacy policy</a>.
 
 # registration-tax-id-itin
