+++ src/main/resources/templates/providerresponse/registration-home-provider-ssn.html
@@ -9,17 +9,9 @@
             <div th:replace="~{fragments/goBack :: goBackLink}"></div>
             <main id="content" role="main" class="form-card spacing-above-35">
                 <th:block th:replace="~{fragments/gcc-icons :: social-security}"></th:block>
-
-                <th:block th:unless="${@environment.getProperty('il-gcc.enable-provider-registration-with-itin') == 'true'}">
-                    <th:block
-                            th:replace="~{fragments/cardHeaderForSingleInputScreen ::
-              cardHeaderForSingleInputScreen(header=#{registration-home-provider-ssn.header}, subtext=#{registration-home-provider-ssn.subtext}, required='true')}"/>
-                </th:block>
-                <th:block th:if="${@environment.getProperty('il-gcc.enable-provider-registration-with-itin') == 'true'}">
-                    <th:block
-                            th:replace="~{fragments/cardHeaderForSingleInputScreen ::
+                <th:block
+                        th:replace="~{fragments/cardHeaderForSingleInputScreen ::
               cardHeaderForSingleInputScreen(header=#{registration-home-provider-ssn.header}, required='true')}"/>
-                </th:block>
 
                 <th:block
                         th:replace="~{fragments/form :: form(action=${formAction}, content=~{::providerLicensed})}">
