+++ src/main/resources/templates/providerresponse/registration-payment-tax-id.html
@@ -1,5 +1,5 @@
 <th:block
-    th:replace="~{fragments/screens/screenWithOneInput ::
+        th:replace="~{fragments/screens/screenWithOneInput ::
   screenWithOneInput(
     title=#{registration-payment-tax-id.title},
     header=#{registration-payment-tax-id.header},
@@ -9,43 +9,39 @@
     buttonLabel=#{general.button.continue},
     required='true',
     inputContent=~{::inputContent})}">
-  <th:block th:ref="inputContent">
-    <div class="spacing-below-60">
-      <th:block th:replace="~{fragments/honeycrisp/reveal :: reveal(
+    <th:block th:ref="inputContent">
+        <div class="spacing-below-60">
+            <th:block th:replace="~{fragments/honeycrisp/reveal :: reveal(
         controlId='r1',
         linkLabel=~{::revealLabel},
         content=~{::revealContent},
         forceShowContent='false')}">
-        <th:block th:ref="revealLabel">
-          <th:span th:text="#{registration-payment-tax-id.reveal-header}"></th:span>
-        </th:block>
-        <th:block th:ref="revealContent">
-          <th:span th:utext="#{registration-payment-tax-id.reveal-body}"></th:span>
-        </th:block>
-      </th:block>
-    </div>
-    <th:block th:replace="~{fragments/inputs/radioFieldset ::
+                <th:block th:ref="revealLabel">
+                    <th:span th:text="#{registration-payment-tax-id.reveal-header}"></th:span>
+                </th:block>
+                <th:block th:ref="revealContent">
+                    <th:span th:utext="#{registration-payment-tax-id.reveal-body}"></th:span>
+                </th:block>
+            </th:block>
+        </div>
+        <th:block th:replace="~{fragments/inputs/radioFieldset ::
                     radioFieldset(inputName='providerTaxIdType',
                     ariaLabel='header',
                     content=~{::radioContent})}">
-      <th:block th:ref="radioContent">
-        <th:block
-            th:with="homeProvider = ${!inputData.getOrDefault('homeProviderTaxIDselection', '').equals('')},
+            <th:block th:ref="radioContent">
+                <th:block
+                        th:with="homeProvider = ${!inputData.getOrDefault('homeProviderTaxIDselection', '').equals('')},
                      hideSSNProviderTypeInput=${homeProvider && (inputData.get('homeProviderTaxIDselection').equals('ITIN'))},
                      hideITINOption = ${homeProvider && (inputData.getOrDefault('homeProviderTaxIDselection', '').equals('SSN'))}">
-          <th:block th:unless="${@environment.getProperty('il-gcc.enable-provider-registration-with-itin') == 'true' && hideSSNProviderTypeInput}">
-            <th:block
-                th:replace="~{fragments/inputs/radio :: radio(inputName='providerTaxIdType',value='SSN', label=#{registration-payment-tax-id.ssn})}"/>
-          </th:block>
-          <th:block
-              th:if="${@environment.getProperty('il-gcc.enable-provider-registration-with-itin') == 'true' && !hideITINOption}">
-            <th:block
-                th:replace="~{fragments/inputs/radio :: radio(inputName='providerTaxIdType',value='ITIN', label=#{registration-payment-tax-id.itin})}"/>
-          </th:block>
-          <th:block
-              th:replace="~{fragments/inputs/radio :: radio(inputName='providerTaxIdType',value='FEIN', label=#{registration-payment-tax-id.fein})}"/>
+                    <th:block th:unless="${hideSSNProviderTypeInput}"
+                              th:replace="~{fragments/inputs/radio :: radio(inputName='providerTaxIdType',value='SSN', label=#{registration-payment-tax-id.ssn})}"/>
+                    <th:block
+                            th:if="${!hideITINOption}"
+                            th:replace="~{fragments/inputs/radio :: radio(inputName='providerTaxIdType',value='ITIN', label=#{registration-payment-tax-id.itin})}"/>
+                    <th:block
+                            th:replace="~{fragments/inputs/radio :: radio(inputName='providerTaxIdType',value='FEIN', label=#{registration-payment-tax-id.fein})}"/>
+                </th:block>
+            </th:block>
         </th:block>
-      </th:block>
     </th:block>
-  </th:block>
 </th:block>
